fix(platform): harden srcdoc app asset routing
